E-Library
E-Library is a Library Management System that allow user to create account, borrow book, and strore it back. This program is created in C language but use the C++ format file and the data was store in a file with .dat extension. The purpose we make this project is to fulfill our final project for Algorithm and Programming courses.